+++ core/sessions.py
@@ -12,15 +12,17 @@ import urlparse
 import atexit
 import ast
 
-print_filters = [
+print_filters = (
     'debug',
-    'channel'
-]
+    'channel',
+    'proxy'
+)
 
-set_filters = [
+set_filters = (
     'debug',
-    'channel'
-]
+    'channel',
+    'proxy'
+)
 
 class Session(dict):
 
@@ -56,6 +58,22 @@ class Session(dict):
          path = self['file_cd']['results'].get('cwd', '.')
      )
 
+    def load_session(self, data):
+        """
+        Update the session dictionary, and occasionally run
+        action_<arg> function.
+        """
+
+        self.update(data)
+
+        for module_argument, value in data.items():
+
+            # If action_<module_argument> function exists, trigger the action
+            action_name = 'action_%s' % (module_argument.replace('.','_'))
+            if hasattr(self, action_name):
+                action_func = getattr(self, action_name)
+                if hasattr(action_func, '__call__'):
+                    action_func(module_argument, value)
 
     def action_debug(self, module_argument, value):
 
@@ -63,6 +81,13 @@ class Session(dict):
             stream_handler.setLevel(logging.DEBUG)
         else:
             stream_handler.setLevel(logging.INFO)
+
+
+    def action_proxy(self, module_argument, value):
+        """After setting a new proxy, reinitiate channels"""
+
+        self['shell_php']['status'] = Status.IDLE
+
 
     def set(self, module_argument, value):
         """Called by user to set or show the session variables"""
@@ -93,7 +118,7 @@ class Session(dict):
                 log.info("%s.%s = '%s'" % (module_name, arg_name, value))
         else:
             module_name = module_argument
-            if module_name not in self or module_name not in set_filters:
+            if module_name not in self and module_name not in set_filters:
                 log.warn(messages.sessions.error_storing_s_not_found % (module_name))
             else:
                 self[module_name] = value
@@ -124,7 +149,7 @@ class SessionFile(Session):
                 # Register dump at exit and return
                 atexit.register(self._session_save_atexit)
 
-            self.update(sessiondb)
+            self.load_session(sessiondb)
             return
 
         log.warn(
@@ -179,7 +204,7 @@ class SessionURL(Session):
                     if not volatile:
                         atexit.register(self._session_save_atexit)
 
-                    self.update(sessiondb)
+                    self.load_session(sessiondb)
                     return
 
         # If no session was found, create a new one with first available filename
@@ -208,7 +233,7 @@ class SessionURL(Session):
                 if not volatile:
                     atexit.register(self._session_save_atexit)
 
-                self.update(sessiondb)
+                self.load_session(sessiondb)
                 return
 
             else: